% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adutils.R
\name{adauth}
\alias{adauth}
\title{AD authentication}
\usage{
adauth(address, user_email, password)
}
\arguments{
\item{address}{LDAP hostname or IP without protocol (ldap://)}

\item{user_email}{user email address}

\item{password}{password}
}
\value{
response indicating whether authentication is successful
}
\description{
\code{adauth} is a function to authenticate against the Active Directory.
}
\details{
\code{user_email} and \code{password} are used to authenticate on \code{address}. Either host name or IP address can be specified.

The response is a list of 2 elements: \code{is_authenticated} and \code{message}. \code{message} is \code{NULL} if \code{is_authenticated} is \code{TRUE}.
}
\examples{
\dontrun{

adauth('hostname-or-ip', 'user_email', 'password')
}
}

