% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radhelper.R
\name{radhelper}
\alias{radhelper}
\title{R wrapper of Python ldap library}
\description{
\code{radhelper} is a wrapper of Python ladp library and is planned to include a set of functions to work on Windows AD directory.
}
\details{
Currently only one function, \code{adauth()}, is implemented.

Python ldap - \url{https://www.python-ldap.org/}

Introduction - \url{https://thomastoye.be/2015/ldap-with-python-and-ad/}
}

